feat(config): update site URLs and metadata for improved SEO and routing
