+++ .env.example
@@ -22,6 +22,7 @@ GEMINI_ALLOW_PRO_MODELS=false
 
 # Frontend
 NEXT_PUBLIC_API_URL=http://localhost:4000/api/v1
+NEXT_PUBLIC_SITE_URL=http://localhost:3000
 
 # MCP Server
 MCP_PORT=5100
