+++ apps/web/next.config.js
@@ -2,6 +2,7 @@
 const nextConfig = {
   transpilePackages: ["@wealthwise/shared-types"],
   output: "standalone",
+  poweredByHeader: false,
 };
 
 module.exports = nextConfig;
