+++ apps/web/public/site.webmanifest
@@ -1,0 +1,27 @@
+{
+  "name": "WealthWise - Smart Personal Finance Tracker",
+  "short_name": "WealthWise",
+  "description": "Track expenses, budgets, goals, recurring bills, and financial analytics in one secure dashboard.",
+  "theme_color": "#0f172a",
+  "background_color": "#ffffff",
+  "display": "standalone",
+  "scope": "/",
+  "start_url": "/",
+  "icons": [
+    {
+      "src": "/android-chrome-192x192.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/android-chrome-512x512.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    },
+    {
+      "src": "/apple-touch-icon.png",
+      "sizes": "180x180",
+      "type": "image/png"
+    }
+  ]
+}
