+++ apps/web/src/app/(auth)/auth-layout-client.tsx
@@ -1,0 +1,64 @@
+"use client";
+
+import { Moon, Sun, Wallet } from "lucide-react";
+import Link from "next/link";
+import { useTheme } from "next-themes";
+import { Button } from "@/components/ui/button";
+import { AppFooter } from "@/components/layout/app-footer";
+
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
+  const { theme, setTheme } = useTheme();
+
+  return (
+    <div className="relative flex min-h-screen flex-col items-center justify-center overflow-hidden px-4 py-12">
+      {/* Background gradients */}
+      <div className="pointer-events-none absolute inset-0">
+        <div className="absolute -right-40 -top-40 h-[500px] w-[500px] rounded-full bg-primary/10 blur-3xl" />
+        <div className="absolute -bottom-40 -left-40 h-[500px] w-[500px] rounded-full bg-violet-500/10 blur-3xl" />
+        <div className="absolute left-1/3 top-1/4 h-[300px] w-[300px] rounded-full bg-emerald-500/5 blur-3xl" />
+      </div>
+
+      {/* Subtle grid pattern */}
+      <div
+        className="pointer-events-none absolute inset-0 opacity-[0.015] dark:opacity-[0.03]"
+        style={{
+          backgroundImage:
+            "linear-gradient(hsl(var(--foreground)) 1px, transparent 1px), linear-gradient(to right, hsl(var(--foreground)) 1px, transparent 1px)",
+          backgroundSize: "60px 60px",
+        }}
+      />
+
+      {/* Theme toggle - top right */}
+      <div className="absolute right-4 top-4 z-20">
+        <Button
+          variant="ghost"
+          size="icon"
+          onClick={() => setTheme(theme === "dark" ? "light" : "dark")}
+        >
+          <Sun className="h-5 w-5 rotate-0 scale-100 text-muted-foreground transition-all dark:-rotate-90 dark:scale-0" />
+          <Moon className="absolute h-5 w-5 rotate-90 scale-0 text-muted-foreground transition-all dark:rotate-0 dark:scale-100" />
+          <span className="sr-only">Toggle theme</span>
+        </Button>
+      </div>
+
+      {/* Logo */}
+      <Link
+        href="/"
+        className="relative z-10 mb-8 flex items-center gap-2.5 transition-opacity hover:opacity-80"
+      >
+        <div className="flex h-10 w-10 items-center justify-center rounded-xl bg-primary text-primary-foreground shadow-lg shadow-primary/25">
+          <Wallet className="h-5 w-5" />
+        </div>
+        <span className="text-2xl font-bold tracking-tight">WealthWise</span>
+      </Link>
+
+      {/* Auth card */}
+      <div className="relative z-10 w-full max-w-md">{children}</div>
+
+      {/* Footer */}
+      <div className="relative z-10 mt-auto w-full pt-8">
+        <AppFooter />
+      </div>
+    </div>
+  );
+}
