+++ apps/web/src/app/(auth)/login/page.tsx
@@ -5,11 +5,20 @@ export const metadata: Metadata = {
   title: "Sign In",
   description:
     "Sign in to your WealthWise account to access your personal finance dashboard, track expenses, and manage your budgets.",
+  alternates: {
+    canonical: "/login",
+  },
   openGraph: {
     title: "Sign In | WealthWise",
     description:
       "Sign in to your WealthWise account to access your personal finance dashboard, track expenses, and manage your budgets.",
-    url: "https://wealthwisefinancial.vercel.app/login",
+    url: "/login",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "Sign In | WealthWise",
+    description:
+      "Sign in to your WealthWise account to access your personal finance dashboard, track expenses, and manage your budgets.",
   },
 };
