+++ apps/web/src/app/(auth)/register/page.tsx
@@ -5,11 +5,20 @@ export const metadata: Metadata = {
   title: "Create Account",
   description:
     "Create your free WealthWise account and start taking control of your personal finances today. Track expenses, set budgets, and reach your savings goals.",
+  alternates: {
+    canonical: "/register",
+  },
   openGraph: {
     title: "Create Account | WealthWise",
     description:
       "Create your free WealthWise account and start taking control of your personal finances today. Track expenses, set budgets, and reach your savings goals.",
-    url: "https://wealthwisefinancial.vercel.app/register",
+    url: "/register",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "Create Account | WealthWise",
+    description:
+      "Create your free WealthWise account and start taking control of your personal finances today. Track expenses, set budgets, and reach your savings goals.",
   },
 };
